refactor: split c2000_cmds.py into modular subpackage (PR review fix)
Addresses modularity review feedback - c2000_cmds.py was 316 lines, exceeding
the 300-line limit. Split into focused modules:

- eab/cli/c2000/__init__.py (13 lines) — package exports
- eab/cli/c2000/reg_cmds.py (154 lines) — reg-read, erad-status
- eab/cli/c2000/stream_cmds.py (122 lines) — stream-vars, dlog-capture
- eab/cli/c2000/trace_cmds.py (83 lines) — c2000-trace-export
- eab/cli/c2000/formatters.py (36 lines) — display formatting helpers

All 25 CLI tests passing. No functionality changed, pure refactor.
